move tiered cache related code to sstable store
